Bugfixes (#13)
* Increase CSS rule specificity

I thought that CSS rules for components were only applied to that
particular component, but I was wrong.  I wrote a rule that wasn't
specific enough.

* Add scale transform to mobile pop-in navbar

I had been hiding the navbar off screen but realized that it was
still being rendered on mobile and resulted in a right scrollbar.
Now it will scale to 0 when not active, but will still be accessible
to screen readers.  (Does anyone use a screen reader with a mobile?)