Define props for Leaderboard
This addresses a warning that would appear in the console:
"currentUser" was being passed via the Vue router but not taken into
use.  The downside is that now we have an unused "props" variable, but
at least this error is hidden from the users.

And perhaps we could add something to the Leaderboard tables whereby
if userName == currentUser, the row is highlighted.